backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiosmtplib==3.0.1
jinja2==3.1.2
pqcrypto==0.3.4
opencv-python==4.8.1.78
numpy==1.26.2
scikit-learn==1.3.2
scikit-image==0.22.0
pillow==10.1.0